Fix travis time out
The `WARNING: could not attach metadata for @simd loop` occurs when
inlining is turned off, which is the case when running `Pkg.test(;
coverage=true)`, so the code is much slower than usual. The solution is
to use a 100x100 matrix as the test example instead of a 500x500 one.